Add stock data source parameter
Add `force_data_source` parameter to Stock.__init__ allowing
manual selection of data source ('twse' or 'tpex'), this also
implied that we will not check if the stock ID exist in the database
or not.

For example, the TPEX has some warrants:

上櫃認購(售)權證,70000U,譜瑞群益39售07
上櫃認購(售)權證,70001U,宣德國票3A售01

which does not exist in the latest codes.

Perhaps someone would like to know their history price,
if they update the codes, twstock will block them down.

But, in this commit, you can explicit passing which data source
you want, and implicitly bypass the stock ID check:

 # Skip stock ID validation (use with caution)
 stock = twstock.Stock('731120', force_data_source="twse")